feat: Implement machine management UI, dashboard, and database stored procedures for production and shift data. Remove MQTT integration

+++ database/stored_procedures.sql
@@ -108,7 +108,7 @@ RETURNS TABLE(
 BEGIN
     RETURN QUERY
     SELECT 
-        ps.machine_id,
+        ps.machine_id::BIGINT,
         m.name AS machine_name,
         m.type::VARCHAR(20) AS machine_type,
         ps.total_units,
